requirements.txt
annotated-types==0.6.0 ; python_version >= "3.11" and python_version < "4.0"
attrs==23.2.0 ; python_version >= "3.11" and python_version < "4.0"
certifi==2024.2.2 ; python_version >= "3.11" and python_version < "4.0"
charset-normalizer==3.3.2 ; python_version >= "3.11" and python_version < "4.0"
dotty-dict==1.3.1 ; python_version >= "3.11" and python_version < "4.0"
flywheel-gear-toolkit==0.6.18 ; python_version >= "3.11" and python_version < "4.0"
flywheel-gears==0.3.1 ; python_version >= "3.11" and python_version < "4.0"
flywheel-sdk==16.19.2 ; python_version >= "3.11" and python_version < "4.0"
fw-client==0.7.5 ; python_version >= "3.11" and python_version < "4.0"
fw-http-client==1.7.2 ; python_version >= "3.11" and python_version < "4.0"
fw-utils==4.4.1 ; python_version >= "3.11" and python_version < "4.0"
idna==3.7 ; python_version >= "3.11" and python_version < "4.0"
jsonschema-specifications==2023.12.1 ; python_version >= "3.11" and python_version < "4.0"
jsonschema==4.21.1 ; python_version >= "3.11" and python_version < "4.0"
numpy==1.26.4 ; python_version >= "3.12" and python_version < "4.0" or python_version == "3.11"
packaging==24.0 ; python_version >= "3.11" and python_version < "4.0"
pandas==2.2.2 ; python_version >= "3.11" and python_version < "4.0"
pydantic-core==2.18.1 ; python_version >= "3.11" and python_version < "4.0"
pydantic==2.7.0 ; python_version >= "3.11" and python_version < "4.0"
python-dateutil==2.9.0.post0 ; python_version >= "3.11" and python_version < "4.0"
pytz==2024.1 ; python_version >= "3.11" and python_version < "4.0"
referencing==0.34.0 ; python_version >= "3.11" and python_version < "4.0"
requests-toolbelt==1.0.0 ; python_version >= "3.11" and python_version < "4.0"
requests==2.31.0 ; python_version >= "3.11" and python_version < "4.0"
rfc3987==1.3.8 ; python_version >= "3.11" and python_version < "4.0"
rpds-py==0.18.0 ; python_version >= "3.11" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.11" and python_version < "4.0"
typing-extensions==4.11.0 ; python_version >= "3.11" and python_version < "4.0"
tzdata==2024.1 ; python_version >= "3.11" and python_version < "4.0"
tzlocal==5.2 ; python_version >= "3.11" and python_version < "4.0"
urllib3==2.2.1 ; python_version >= "3.11" and python_version < "4.0"
